% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wd.R
\name{cd}
\alias{cd}
\title{Change (Working) Directory}
\usage{
cd(loc)
}
\arguments{
\item{loc}{where you want to set the working directory}
}
\description{
This function allows setwd() and getwd() in one step.
}
\examples{
cd("~/Files")
}
\keyword{directory}
\keyword{wd,}
\keyword{working}
