% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdC.R
\name{cmdC}
\alias{cmdC}
\title{PBCopy}
\usage{
cmdC(x, clipboardname = "pbcopy", sep = " ")
}
\description{
Write to the clipboard. Useful for exporting random data, especially in Excel.
}
\examples{
cmdC()
}
\keyword{paste,}
\keyword{table}
