% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdVtable.R
\name{cmdVtable}
\alias{cmdVtable}
\title{PBPaste Table}
\usage{
cmdVtable(clipboardname = "pbpaste", header = FALSE, sep = "",
  quote = "\\"'", dec = ".")
}
\description{
Read table from the clipboard.
}
\examples{
cmdVtable()
}
\keyword{paste,}
\keyword{table}
