% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctrlC.R
\name{ctrlC}
\alias{ctrlC}
\title{PBCopy}
\usage{
ctrlC(x, clipboardname = "pbcopy", sep = " ")
}
\description{
Write to the clipboard. Useful for exporting random data, especially in Excel.
}
\examples{
ctrlC()
}
\keyword{paste,}
\keyword{table}
