% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctrlVtable.R
\name{ctrlVtable}
\alias{ctrlVtable}
\title{PBPaste Table}
\usage{
ctrlVtable(clipboardname = "pbpaste", header = FALSE, sep = "",
  quote = "\\"'", dec = ".")
}
\description{
Paste table from the clipboard.
}
\examples{
ctrlVtable()
}
\keyword{paste,}
\keyword{table}
