% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google.R
\name{google}
\alias{google}
\title{Run Google Search}
\usage{
google(query)
}
\arguments{
\item{query}{your search}
}
\description{
Run Google Search from string
}
\examples{
google('how to do something in R')
}
\keyword{google,}
\keyword{search}
