% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headdf.R
\name{headdf}
\alias{headdf}
\title{Data.frame head}
\usage{
headdf(data)
}
\arguments{
\item{data}{a dataset you'd like to use.}
}
\description{
This function allows coerce head data to a dataframe. Useful for something, I just forget what.
}
\examples{
headdf()
}
\keyword{df}
