% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.check.R
\name{library.check}
\alias{library.check}
\title{Load Library or Install Package}
\usage{
library.check(pkg)
}
\arguments{
\item{package_name}{the name of the package you want to load}
}
\description{
Load Library if it exists or install.packages if not
}
\examples{
packages<-c("ggplot2", "afex", "ez", "Hmisc", "pander", "plyr")
check.packages(packages)
}
\keyword{wd}
