% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sopen.R
\name{sopen}
\alias{sopen}
\title{Open}
\usage{
sopen(string)
}
\arguments{
\item{string}{What you want to open from the shell}
}
\description{
Open Shell Command
}
\examples{
sopen("~/Files/")
#' sopen("http://google.com/")
}
\keyword{open}
