% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.Toolshed-package.R
\docType{package}
\name{R.Toolshed-package}
\alias{R.Toolshed}
\alias{R.Toolshed-package}
\title{R.Toolshed: Miscellaneous R Tools}
\description{
Miscellaneous R Tools for easier coding.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/chrisvacc/R.Toolshed}
  \item Report bugs at \url{https://github.com/chrisvacc/R.Toolshed/issues}
}

}
\author{
\strong{Maintainer}: Chris Vaccaro \email{chrisvacc@gmail.com}

}
\keyword{internal}
